Numeros enteros

+++ #Sumar.py
@@ -1,7 +1,7 @@
 #Sumar
 
-num1=float(input("Ingrese numero 1: "))
-num2=float(input("Ingrese numero 2: "))
+num1=int(input("Ingrese numero 1: "))
+num2=int(input("Ingrese numero 2: "))
 
 sumar=num1+num2
-print("La suma de los dos numeros es:",sumar )
+print("La suma de los dos numeros es:",sumar)
